[PATCH] x86-64 update for 2.5.62-bk4
This still has a few problems; e.g. posix-timers.c seems to be miscompiled
by many x86-64 compilers when compiled with -O1. But overall it is better
than before.

 - Disable numa again that was enabled by mistake
 - Fix CPU detection in early boot. Force SSE when needed.
 - Hack for nested NMIs.
 - Fix typo in vsyscall code that lead to broken time reporting
 - Merge with 2.5.62-bk4 (new switch_to, printk cleanups, new timer calls etc.)
 - Fix bug in strace - don't assume syscall_trace doesn't clobber %rcx
 - Fix gs load workaround
 - Fix return value of ioperm (found by new LTP test)
 - Remove some unecessary boot output
 - Remove some obsolete code in setup.c
 - Report more in /proc/cpuinfo, fix extended cpu family computation
 - Fix thread_info access from interrupt/exception stacks - this
   fixes many problems that especially showed on SMP (lost reschedule
   interrupts etc.)
 - Fix UP compilation
 - Symlink hugetlbpage.c from i386
 - Fix buggy lazy fpu handling

+++ arch/x86_64/kernel/ioport.c
@@ -53,7 +53,7 @@ static void set_bitmap(unsigned long *bitmap, short base, short extent, int new_
 /*
  * this changes the io permissions bitmap in the current task.
  */
-asmlinkage int sys_ioperm(unsigned long from, unsigned long num, int turn_on)
+asmlinkage long sys_ioperm(unsigned long from, unsigned long num, int turn_on)
 {
 	struct thread_struct * t = &current->thread;
 	struct tss_struct * tss;
